requirements.txt
streamlit
g4f
mysql-connector-python
cvxpy
tensorflow
tensorflow-hub
numpy
pandas
